Fixed the failing test re points scored after a win
However, unearthed a bug in the tests re scores like 5/4. I am currently
scoring 5 points for player 1 followed by 4 points for player 2. Now
that I have fixed the failing test, this makes tests like 5/4 fail
because the game is over after scoring 4 points for player 1. So, I need
to fix these tests. The most flexible way to do this is to feed a string
like 1,1,1,2,2,1,1... to the test to indicate the order in which points
are won.